nat: split source.rs into a module directory (pure code motion)
`userspace-dp/src/nat/source.rs` crossed the 2000-line [REFACTOR] tier
in #6815 at 2035 lines and has since reached 3315. Six clusters move
into `source/{failure,expand,release,synced,nat64_ports,match_rules}.rs`,
leaving `source/mod.rs` at 1367 — below the 1500 [WATCH] floor.

The clusters were chosen by dependency COST rather than by name. The
governing number is lines moved per visibility widening required, and it
separates the candidates sharply:

  match_rules   747 lines   0 widenings
  synced        515 lines   0
  release       257 lines   0
  expand        174 lines   0 (one respelling — see below)
  failure       157 lines   2
  nat64_ports   140 lines   0

Two widenings for 1990 lines moved, both in failure.rs:
`SourceNatFailure::for_rule` (19 callers, all in match_rules) and
`source_nat_failure_reason_from_snapshot` (one caller, in the parse
pipeline that stays). Nothing else needed one, because a PRIVATE item is
visible to its own module AND that module's descendants — so everything
that stayed in mod.rs and is called from a submodule
(`SourceNatRule::matches`, `matches_ignoring_scope`,
`address_has_draining_pool_occupancy`, `parse_match_prefix`,
`port_in_ranges`, `nets_match_v4/v6`) moved to zero cost.

DECLINED, and recorded because it is the seam the issue named first: the
pending-allocator resolution pass (`PreviousPool` /
`PendingPoolAllocator` / `resolve_pool_allocators`, ~360 lines). Moving
it costs SEVENTEEN widenings, including making two private structs and
all six of their fields `pub(super)` because
`parse_source_nat_rules_inner` CONSTRUCTS them. That is not a seam — it
is one property scattered across two files — so it stays with the parse
pipeline it belongs to, together with the aggregate-budget cluster that
shares `SourceNatAggregateUse::saturating_with` with it.

The one non-obvious edit is a visibility PRESERVATION, not a change.
`expand_pool_address` was `pub(super)` inside `nat::source`, meaning
visible in `nat`, which is how `tests_aggregate_budget.rs` reaches it.
One level deeper the same keyword means `pub(in crate::nat::source)` —
narrower — and it broke that caller with E0603 on the first build. It is
now spelled `pub(in crate::nat)`: exactly the visibility the pre-split
file had. The caution is recorded in `nat/mod.rs` because the next
person to move a `pub(super)` item deeper will hit it too.

Nothing was fixed while moving. Two dead-ish items travelled untouched:
`SourceNatRule::allocator_key` (no callers anywhere, no
`#[allow(dead_code)]`, warns at master too) and
`source_nat_runtime_compatible` (explicitly allowed). Fixing either
inside a 1990-line motion diff would be invisible to a reviewer.

Two machine checks stand in for a reviewer reading 1990 moved lines:

  1. `scripts/verify-nat-source-split-6988.py` reconstructs the
     pre-split `source.rs` from the split and compares it byte-for-byte
     with `origin/master`, allowing ONLY the three enumerated edits. It
     PASSES at 3316 lines. A reordered statement, a changed literal or a
     dropped comment fails it.
  2. The compiler's own diagnostics are identical: 168 warnings at
     master, 168 here, with zero added and zero removed (diffed as sets
     with the source path normalised). A motion that changed behaviour
     would be very unlikely to leave that set untouched.

Validation: `cargo test --release --bins -- --test-threads=1` — 4745
collected, 4743 passed / 0 failed / 2 ignored; `cargo build --release`
clean; `go test ./pkg/refactoraudit/` ok.
